+++ inst/staticdocs/index.R
@@ -4,6 +4,7 @@ sd_section("Templating functions",
   "Functions for constructing package and file templates.",
   c(
       "pax",
+      "new_r_test",
       "new_r",
       "new_test"
   )
@@ -14,6 +15,7 @@ sd_section("General use function",
   "Functions intended for internal or general use.",
   c(
       "is.global",
+      "pax_options",
       "ploc"
   )
 )